fix: comprehensive binding overhaul in SettingsWindow
Root cause: multiple WPF binding mode and type mismatches.

Fixes applied:
- DownloadProgress: add Mode=OneWay (ProgressBar.Value defaults to
  TwoWay, but property has private setter → crash)
- DownloadStatus: add Mode=OneWay (same reason)
- VadThresholdDb: change from float to double in ViewModel
  (Slider.Value is double, TwoWay binding double→float crashes)
- VadSilenceMs: change from int to double in ViewModel (same reason)
- Remove all Run.Text bindings (unreliable DataContext inheritance),
  replace with TextBlock.Text + StringFormat
- Add DisplayMemberPath to Languages ComboBox
- Cast double→float/int only in Save() when writing back to settings

+++ src/Voicy/ViewModels/SettingsViewModel.cs
@@ -104,16 +104,16 @@ public class SettingsViewModel : ViewModelBase
         set => SetProperty(ref _isCapturingHotkey, value);
     }
 
-    // VAD
-    private float _vadThresholdDb = -30f;
-    public float VadThresholdDb
+    // VAD — use double to match Slider.Value type (avoids TwoWay binding type mismatch)
+    private double _vadThresholdDb = -30.0;
+    public double VadThresholdDb
     {
         get => _vadThresholdDb;
         set => SetProperty(ref _vadThresholdDb, value);
     }
 
-    private int _vadSilenceMs = 800;
-    public int VadSilenceMs
+    private double _vadSilenceMs = 800.0;
+    public double VadSilenceMs
     {
         get => _vadSilenceMs;
         set => SetProperty(ref _vadSilenceMs, value);
@@ -197,8 +197,8 @@ public class SettingsViewModel : ViewModelBase
         _settings.Mode = Mode;
         _settings.HotkeyModifiers = _hotkeyModifiers;
         _settings.HotkeyKey = _hotkeyKey;
-        _settings.VadThresholdDb = VadThresholdDb;
-        _settings.VadSilenceMs = VadSilenceMs;
+        _settings.VadThresholdDb = (float)VadThresholdDb;
+        _settings.VadSilenceMs = (int)VadSilenceMs;
 
         _settingsService.Save(_settings);
         Saved = true;
